requirements.txt
langchain
langchain-community
langchain-openai
langgraph
tavily-python
chromadb
sentence-transformers
rank-bm25
pypdf
python-dotenv
streamlit
openai
numpy